finish user module

+++ image-hub-backend/src/main/java/com/ziio/imagehubbackend/common/PageRequest.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class PageRequest {
 
-    private int current = 1;
+    private long current = 1;
 
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     private String sortField;
